refactor : Contract 오류로 인한 코드 수정
