+++ src/main/java/fourservings_fiveservings/insurance_system_be/team/contract/entity/common/ApproveStatus.java
@@ -1,0 +1,7 @@
+package fourservings_fiveservings.insurance_system_be.team.contract.entity.common;
+
+public enum ApproveStatus {
+    APPROVE,
+    UN_APPROVE,
+    PENDING
+}
