+++ src/main/java/fourservings_fiveservings/insurance_system_be/team/contract/repository/ContractRepository.java
@@ -1,7 +1,7 @@
 package fourservings_fiveservings.insurance_system_be.team.contract.repository;
 
+import fourservings_fiveservings.insurance_system_be.team.contract.entity.common.ApproveStatus;
 import fourservings_fiveservings.insurance_system_be.team.contract.entity.common.Contract;
-import fourservings_fiveservings.insurance_system_be.team.plan.insurance.ApproveStatus;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
