+++ src/main/java/fourservings_fiveservings/insurance_system_be/team/insurance/entity/SaleTarget.java
@@ -1,0 +1,7 @@
+package fourservings_fiveservings.insurance_system_be.team.insurance.entity;
+
+public enum SaleTarget {
+    YOUNG,
+    MIDDLE_AGED,
+    OLDER
+}
