+++ src/main/java/fourservings_fiveservings/insurance_system_be/team/insurance/implement/InsuranceFinder.java
@@ -1,0 +1,22 @@
+package fourservings_fiveservings.insurance_system_be.team.insurance.implement;
+
+import fourservings_fiveservings.insurance_system_be.common.exception.ResourceNotFoundException;
+import fourservings_fiveservings.insurance_system_be.common.exception.constant.ErrorType;
+import fourservings_fiveservings.insurance_system_be.team.insurance.entity.Insurance;
+import fourservings_fiveservings.insurance_system_be.team.insurance.repository.InsuranceRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class InsuranceFinder {
+
+    private final InsuranceRepository insuranceRepository;
+
+    public Insurance findById(Long id) {
+        return insuranceRepository.findById(id)
+            .orElseThrow(() -> new ResourceNotFoundException(ErrorType.RESOURCE_NOT_FOUND));
+    }
+}
